requirements.txt
certifi==2018.1.18
chardet==3.0.4
cssselect==1.0.3
diff-match-patch==20121119
Django==1.11.6
django-allauth==0.33.0
django-import-export==1.0.0
django-ipware==1.1.6
django-js-asset==0.1.1
et-xmlfile==1.0.1
idna==2.6
jdcal==1.3
Jinja2==2.10
lxml==4.0.0
MarkupSafe==1.0
mysqlclient==1.3.7
oauthlib==2.0.6
odfpy==1.3.6
olefile==0.44
openpyxl==2.5.0
parsel==1.2.0
Pillow==4.1.1
pyasn1==0.3.6
pyasn1-modules==0.1.4
pycparser==2.18
PyDispatcher==2.0.5
python-openid==2.2.5
pytz==2018.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
sqlparse==0.1.18
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
virtualenv==15.1.0
w3lib==1.19.0
xlrd==1.1.0
xlwt==1.3.0
